use macro for wasm binding implementation for `PodElGamalPubkey`
