+++ zk-sdk/src/encryption/pod/elgamal.rs
@@ -11,7 +11,7 @@ use {
 use {
     crate::{
         encryption::{DECRYPT_HANDLE_LEN, ELGAMAL_CIPHERTEXT_LEN, ELGAMAL_PUBKEY_LEN},
-        pod::{impl_from_bytes, impl_from_str},
+        pod::{impl_from_bytes, impl_from_str, impl_wasm_bindings},
     },
     base64::{prelude::BASE64_STANDARD, Engine},
     bytemuck::Zeroable,
@@ -88,77 +88,7 @@ impl TryFrom<PodElGamalCiphertext> for ElGamalCiphertext {
 #[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
 pub struct PodElGamalPubkey(pub(crate) [u8; ELGAMAL_PUBKEY_LEN]);
 
-#[cfg(target_arch = "wasm32")]
-#[allow(non_snake_case)]
-#[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
-impl PodElGamalPubkey {
-    /// Create a new `PodElGamalPubkey` object
-    ///
-    /// * `value` - optional public key as a base64 encoded string, `Uint8Array`, `[number]`
-    #[wasm_bindgen(constructor)]
-    pub fn constructor(value: JsValue) -> Result<PodElGamalPubkey, JsValue> {
-        if let Some(base64_str) = value.as_string() {
-            base64_str
-                .parse::<PodElGamalPubkey>()
-                .map_err(|e| e.to_string().into())
-        } else if let Some(uint8_array) = value.dyn_ref::<Uint8Array>() {
-            bytemuck::try_from_bytes(&uint8_array.to_vec())
-                .map_err(|err| JsValue::from(format!("Invalid Uint8Array ElGamalPubkey: {err:?}")))
-                .map(|pubkey| *pubkey)
-        } else if let Some(array) = value.dyn_ref::<Array>() {
-            let mut bytes = vec![];
-            let iterator = js_sys::try_iter(&array.values())?.expect("array to be iterable");
-            for x in iterator {
-                let x = x?;
-
-                if let Some(n) = x.as_f64() {
-                    if (0. ..=255.).contains(&n) {
-                        bytes.push(n as u8);
-                        continue;
-                    }
-                }
-                return Err(format!("Invalid array argument: {:?}", x).into());
-            }
-
-            bytemuck::try_from_bytes(&bytes)
-                .map_err(|err| JsValue::from(format!("Invalid Array pubkey: {err:?}")))
-                .map(|pubkey| *pubkey)
-        } else if value.is_undefined() {
-            Ok(PodElGamalPubkey::default())
-        } else {
-            Err("Unsupported argument".into())
-        }
-    }
-
-    /// Return the base64 string representation of the public key
-    pub fn toString(&self) -> String {
-        self.to_string()
-    }
-
-    /// Checks if two `ElGamalPubkey`s are equal
-    pub fn equals(&self, other: &PodElGamalPubkey) -> bool {
-        self == other
-    }
-
-    /// Return the `Uint8Array` representation of the public key
-    pub fn toBytes(&self) -> Box<[u8]> {
-        self.0.into()
-    }
-
-    pub fn zeroed() -> Self {
-        Self::default()
-    }
-
-    pub fn compressed(decoded: &ElGamalPubkey) -> PodElGamalPubkey {
-        (*decoded).into()
-    }
-
-    pub fn decompressed(&self) -> Result<ElGamalPubkey, JsValue> {
-        (*self)
-            .try_into()
-            .map_err(|err| JsValue::from(format!("Invalid ElGamalPubkey: {err:?}")))
-    }
-}
+impl_wasm_bindings!(POD_TYPE = PodElGamalPubkey, DECODED_TYPE = ElGamalPubkey);
 
 impl fmt::Debug for PodElGamalPubkey {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
