+++ zk-sdk/src/pod.rs
@@ -62,3 +62,75 @@ macro_rules! impl_from_bytes {
     };
 }
 pub(crate) use impl_from_bytes;
+
+macro_rules! impl_wasm_bindings {
+    (POD_TYPE = $pod_type:ident, DECODED_TYPE = $decoded_type: ident) => {
+        #[cfg(target_arch = "wasm32")]
+        #[allow(non_snake_case)]
+        #[cfg_attr(target_arch = "wasm32", wasm_bindgen)]
+        impl $pod_type {
+            #[wasm_bindgen(constructor)]
+            pub fn constructor(value: JsValue) -> Result<$pod_type, JsValue> {
+                if let Some(base64_str) = value.as_string() {
+                    base64_str
+                        .parse::<$pod_type>()
+                        .map_err(|e| e.to_string().into())
+                } else if let Some(uint8_array) = value.dyn_ref::<Uint8Array>() {
+                    bytemuck::try_from_bytes(&uint8_array.to_vec())
+                        .map_err(|err| JsValue::from(format!("Invalid Uint8Array: {err:?}")))
+                        .map(|value| *value)
+                } else if let Some(array) = value.dyn_ref::<Array>() {
+                    let mut bytes = vec![];
+                    let iterator =
+                        js_sys::try_iter(&array.values())?.expect("array to be iterable");
+                    for x in iterator {
+                        let x = x?;
+
+                        if let Some(n) = x.as_f64() {
+                            if (0. ..=255.).contains(&n) {
+                                bytes.push(n as u8);
+                                continue;
+                            }
+                        }
+                        return Err(format!("Invalid array argument: {:?}", x).into());
+                    }
+
+                    bytemuck::try_from_bytes(&bytes)
+                        .map_err(|err| JsValue::from(format!("Invalid Array: {err:?}")))
+                        .map(|value| *value)
+                } else if value.is_undefined() {
+                    Ok($pod_type::default())
+                } else {
+                    Err("Unsupported argument".into())
+                }
+            }
+
+            pub fn toString(&self) -> String {
+                self.to_string()
+            }
+
+            pub fn equals(&self, other: &$pod_type) -> bool {
+                self == other
+            }
+
+            pub fn toBytes(&self) -> Box<[u8]> {
+                self.0.into()
+            }
+
+            pub fn zeroed() -> Self {
+                Self::default()
+            }
+
+            pub fn encode(decoded: &$decoded_type) -> $pod_type {
+                (*decoded).into()
+            }
+
+            pub fn decode(&self) -> Result<$decoded_type, JsValue> {
+                (*self)
+                    .try_into()
+                    .map_err(|err| JsValue::from(format!("Invalid encoding: {err:?}")))
+            }
+        }
+    };
+}
+pub(crate) use impl_wasm_bindings;
